Rename mmd_tools to tools
