+++ dmci/tools/__init__.py
@@ -21,7 +21,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from dmci.mmd_tools.check_mmd import CheckMMD
+from dmci.tools.check_mmd import CheckMMD
 
 __all__ = [
     "CheckMMD",
