+++ pytest.ini
@@ -3,5 +3,5 @@ markers =
     core: Core functionality tests
     dist: Distributor tests
     api: Api functionality tests
-    mmd_tools: MMD toolbox tests
+    tools: Toolbox tests
     serial
